+++ changelog.md
@@ -1,9 +1,10 @@
-# shellbryson.com version
+# shellbryson.com changelog
 
 ## 1.0.0 - 2017-09-26 (initial version)
 
 ### Adds
 
+- Autoreloading via BrowserSync
 - Compile via Grunt.
 - Package information.
 - Breakpoint mixin.